# api-documentation.yaml
swagger: "2.0"
info:
  description: "This is a sample PLG Web server."
  version: 0.0.1
  title: "PLG Web"
  contact:
    email: "andbur@dtu.dk"
host: "localhost:8080"
basePath: "/api/v2/"
schemes:
  - "http"
paths:
...
  /streamer/status/{processName}:
    get:
      tags:
      - "streamer"
      summary: "Checks if a process is being streamed"
      description: ""
      produces:
      - "application/json"
      parameters:
      - in: "path"
        name: "processName"
        type: "string"
        description: "The id of the process to be verified for streaming"
        required: true
      responses:
        "200":
          description: "Whether the process is being streamed or not"
          schema:
            type: "boolean"
  /streamer/start/{processName}:
    post:
      tags:
      - "streamer"
      summary: "Starts streaming the given process"
      description: ""
      produces:
      - "application/json"
      parameters:
      - in: "path"
        name: "processName"
        type: "string"
        description: ""
        required: true
      - in: "body"
        name: "plg"
        description: "The process specification"
        required: true
        schema:
          $ref: "#/definitions/Process4Web"
      - in: "query"
        name: "broker"
        type: "string"
        format: "hostname"
        description: "The host of the broker where to publish the events"
        required: true
      - in: "query"
        name: "topic"
        type: "string"
        description: "The base topic name"
        required: true
      responses:
        "200":
          description: "Whether the stream started to be streamed or not"
          schema:
            type: "boolean"
definitions:
  Process4Web:
    type: "object"
    properties:
      id:
        type: "string"
      name:
        type: "string"
        description: "A short name for the process"
      serialization:
        type: "string"
        description: "A string serialization of the process (can the XML generated from PLG)."
      serializationType:
        type: "string"
        description: "The current serialization type."
        enum:
          - "plg"